メンタリング指摘事項修正２ formImage useStateを使わず、inputItemコンポーネントで状態を保持。 yupバリテーションを変数化。
